ceph.spec.in: remove hard versioned dependency of ceph-test on ceph-common
OBS builds the ceph-test RPM separately from the normal Ceph RPMs, and
ceph-test is shipped on the Internal Media. I also found that OBS has no
problem producing ceph-test RPM with a slightly different version number than
the Ceph RPMs, with the result that it cannot be installed due to this hard
versioned dependency in the spec file.

Upstream doesn't have this problem because they always build ceph-test along
with the other RPMs.

We don't really need the hard versioned dependency. It's convenient to have,
but not absolutely necessary. I would rather add version sanity-checking code
to the test cases that use the ceph-test RPM, than to have all test cases
fail because ceph-test cannot be installed.

Signed-off-by: Nathan Cutler <ncutler@suse.com>