qa/tasks/deepsea.py: rebuild_node fix
This is a fix to ECP CI failures.
Added a sleep in order to prevent the test from failing due to not
rebalancing soon enough plus a test to check if the osds after the
rebuild are equal to the ones before.

Signed-off-by: Georgios Kyratsas <gkyratsas@suse.com>
(cherry picked from commit 1891af4794dec628b6fc9d890388034a67ad6f0e)